pyproject.toml
[project]
name = "line-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.118.3",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
